Spell the 64-bit C API types uint64_t
getBVUnsignedLongLong and vc_bvConstExprFromLL are declared with
unsigned long long int, which says how the type was once spelled rather
than what it means. Both are 64-bit quantities; say so.

The ABI does not move. Both are extern "C", so the exported symbols are
unmangled and unchanged. On i386 and on Windows the edit is a no-op,
because uint64_t is unsigned long long there. On LP64 the spelling
becomes unsigned long, which is the same 64 bits passed the same way --
checked by linking a uint64_t-declared caller against a libstp built
with the old spelling, and an old-spelling caller against a libstp built
with the new one: every boundary value from 0 to UINT64_MAX round-trips
either way.

The Python binding has always declared both as c_uint64, so the header
had been the less precise of the two. It now agrees with the binding it
ships alongside.

stdint.h joins stdbool.h at the top. The header still compiles clean as
C99, C11 and C++17 with -Wall -Wextra -pedantic.

The names stay. Renaming getBVUnsignedLongLong would be a real break for
callers rather than a change of spelling, and it is not worth a
deprecation cycle to lose the "LongLong" in a name that has always meant
64-bit. The one caller this can affect is code taking the address of
either function into an explicitly-typed function pointer: a hard error
in C++, a warning in C. Ordinary calls are unaffected.

+++ lib/Interface/c_interface.cpp
@@ -1176,7 +1176,7 @@ Expr vc_bvConstExprFromInt(VC vc, int n_bits, unsigned int value)
   return persistNode(vc, n);
 }
 
-Expr vc_bvConstExprFromLL(VC vc, int n_bits, unsigned long long value)
+Expr vc_bvConstExprFromLL(VC vc, int n_bits, uint64_t value)
 {
   stp::STP* stp_i = (stp::STP*)vc;
   stp::STPMgr* b = stp_i->bm;
@@ -1803,8 +1803,8 @@ unsigned int getBVUnsigned(Expr e)
   return (unsigned int)a->GetUnsignedConst();
 }
 
-//! Return an unsigned long long int from a constant bitvector expression
-unsigned long long int getBVUnsignedLongLong(Expr e)
+//! Return a uint64_t from a constant bitvector expression
+uint64_t getBVUnsignedLongLong(Expr e)
 {
   stp::ASTNode* a = (stp::ASTNode*)e;
 
@@ -1815,7 +1815,7 @@ unsigned long long int getBVUnsignedLongLong(Expr e)
   unsigned* bv = a->GetBVConst();
 
   char* str_bv = (char*)CONSTANTBV::BitVector_to_Bin(bv);
-  unsigned long long int tmp = std::strtoull(str_bv, NULL, 2);
+  uint64_t tmp = std::strtoull(str_bv, NULL, 2);
   CONSTANTBV::BitVector_Dispose((unsigned char*)str_bv);
   return tmp;
 }
